Move CacheOptions into Options
